+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+### Bug Fixes
+
+-   Fixed packages mode for packages outside of cwd (again), #2043.
+
 ## v0.23.12 (2022-08-31)
 
 ### Features